Add post remove validation. Fix post remove parent node null bug

+++ WebSchool/Controllers/ApiControllers/ApiPostsController.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using System.Security.Claims;
 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
@@ -11,15 +12,24 @@ namespace WebSchool.WebApplication.Controllers.ApiControllers
     public class ApiPostsController : Controller
     {
         private readonly IPostsService postsService;
+        private readonly IUsersService usersService;
 
         public ApiPostsController(
-            IPostsService postsService)
+            IPostsService postsService,
+            IUsersService usersService)
         {
             this.postsService = postsService;
+            this.usersService = usersService;
         }
 
         public async Task<IActionResult> Remove(string postId)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if(usersService.ValidatePostRemove(userId, postId) == false)
+            {
+                return Json(false);
+            }
+
             var removeResult = await postsService.RemoveAsync(postId);
             return Json(removeResult);
         }
